+++ routes/bootcampRoutes.js
@@ -13,12 +13,18 @@ const {
 const courseRouter = require('./courseRoutes');
 const router = express.Router();
 
+const advancedResults = require('../middleware/advancedResults');
+const Bootcamp = require('../models/bootcampModel');
+
 // Re-route into other resource routers
 router.use('/:bootcampId/courses', courseRouter);
 
 router.route('/radius/:zipcode/:distance').get(getBootcampsInRadius);
 
-router.route('/').get(getAllBootcamps).post(addBootcamps);
+router
+  .route('/')
+  .get(advancedResults(Bootcamp, 'courses'), getAllBootcamps)
+  .post(addBootcamps);
 router
   .route('/:id')
   .get(findBootcamp)
